move tests into own class and use a command to enable configuration test
