+++ fabric-networking-api-v1/src/testmod/java/net/fabricmc/fabric/test/networking/play/NetworkingPlayPacketTest.java
@@ -29,7 +29,6 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
 
 import net.minecraft.dialog.type.Dialog;
-import net.minecraft.nbt.NbtElement;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.network.PacketCallbacks;
 import net.minecraft.network.RegistryByteBuf;
@@ -39,7 +38,6 @@ import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.network.packet.s2c.play.BundleS2CPacket;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.registry.RegistryKeys;
-import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
@@ -50,7 +48,6 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
-import net.fabricmc.fabric.api.networking.v1.CustomClickActionEvents;
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
@@ -60,8 +57,6 @@ import net.fabricmc.fabric.test.networking.common.NetworkingCommonTest;
 import net.fabricmc.loader.api.FabricLoader;
 
 public final class NetworkingPlayPacketTest implements ModInitializer {
-	private static final RegistryKey<Dialog> PLAY_TEST_DIALOG = RegistryKey.of(RegistryKeys.DIALOG, NetworkingTestmods.id("play_custom_click_event"));
-
 	private static boolean spamUnknownPackets = false;
 
 	public static void sendToTestChannel(ServerPlayerEntity player, String stuff) {
@@ -108,19 +103,6 @@ public final class NetworkingPlayPacketTest implements ModInitializer {
 					ServerPlayNetworking.reconfigure(ctx.getSource().getPlayer());
 					return Command.SINGLE_SUCCESS;
 				}))
-				.then(literal("testdialog").executes(ctx -> {
-					ServerPlayerEntity player = ctx.getSource().getPlayer();
-
-					if (player != null) {
-						RegistryEntry<Dialog> testDialog = ctx.getSource()
-								.getRegistryManager()
-								.getOrThrow(RegistryKeys.DIALOG)
-								.getOrThrow(PLAY_TEST_DIALOG);
-						player.openDialog(testDialog);
-					}
-
-					return Command.SINGLE_SUCCESS;
-				}))
 		);
 	}
 
@@ -152,15 +134,6 @@ public final class NetworkingPlayPacketTest implements ModInitializer {
 				}
 			}
 		});
-
-		CustomClickActionEvents.playClickActionEvent(NetworkingTestmods.id("play_event")).register(
-				context -> {
-					NbtElement payload = context.payload();
-					String payloadString = payload != null ? payload.toString() : "null";
-					Text message = Text.translatable("key.fabric-networking-api-v1-testmod.customClick.play.received", payloadString);
-					context.handler().getPlayer().sendMessage(message);
-				}
-		);
 	}
 
 	public record OverlayPacket(Text message) implements CustomPayload {
