+++ fabric-networking-api-v1/src/testmod/resources/fabric.mod.json
@@ -15,7 +15,8 @@
       "net.fabricmc.fabric.test.networking.configuration.NetworkingConfigurationTest",
       "net.fabricmc.fabric.test.networking.keybindreciever.NetworkingKeybindPacketTest",
       "net.fabricmc.fabric.test.networking.login.NetworkingLoginQueryTest",
-      "net.fabricmc.fabric.test.networking.play.NetworkingPlayPacketTest"
+      "net.fabricmc.fabric.test.networking.play.NetworkingPlayPacketTest",
+      "net.fabricmc.fabric.test.networking.clickeventtest.CustomClickActionsTest"
     ],
     "client": [
       "net.fabricmc.fabric.test.networking.client.channeltest.NetworkingChannelClientTest",
