+++ fabric-networking-api-v1/src/testmodClient/java/net/fabricmc/fabric/test/networking/client/configuration/NetworkingConfigurationClientTest.java
@@ -16,18 +16,12 @@
 
 package net.fabricmc.fabric.test.networking.client.configuration;
 
-import java.util.Optional;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import net.minecraft.nbt.NbtString;
-import net.minecraft.network.packet.c2s.common.CustomClickActionC2SPacket;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationConnectionEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationNetworking;
-import net.fabricmc.fabric.test.networking.NetworkingTestmods;
 import net.fabricmc.fabric.test.networking.configuration.NetworkingConfigurationTest;
 
 public class NetworkingConfigurationClientTest implements ClientModInitializer {
@@ -51,19 +45,6 @@ public class NetworkingConfigurationClientTest implements ClientModInitializer {
 
 			LOGGER.info("Sending configuration start packet to server");
 			ClientConfigurationNetworking.send(NetworkingConfigurationTest.ConfigurationStartPacket.INSTANCE);
-		});
-
-		// This is a test of the CONFIGURATION event in CustomClickActionEvents. When this packet is received, a message
-		// should appear in the log showing the event has been received by the server, with the payload.
-		// Rather than using an actual dialog (which breaks E2E client tests), this mocks the behaviour of clicking on a
-		// custom event button by sending the packet that such an action would normally send.
-		ClientConfigurationNetworking.registerGlobalReceiver(NetworkingConfigurationTest.MockShowDialogPacket.ID, (packet, context) -> {
-			context.responseSender().sendPacket(
-					new CustomClickActionC2SPacket(
-							NetworkingTestmods.id("configuration_event"),
-							Optional.of(NbtString.of("this is a payload"))
-					)
-			);
 		});
 	}
 }
